third_party/WebKit/LayoutTests/editing/execCommand/insert_horizontal_rule.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable>foo|bar</div>',
    'insertHorizontalRule',
    '<div contenteditable>foo<hr>|bar</div>'),
    'Insert HR between foo and bar');

test(() => assert_selection(
    '<div contenteditable>foo|bar</div>',
    'insertHorizontalRule abc',
    '<div contenteditable>foo<hr id="abc">|bar</div>'),
    'Insert HR with id "abc" between foo and bar');
</script>
